--- prisma/seed.ts
@@ -1,24 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-const prisma = new PrismaClient()
-
-async function main() {
-    const user = await prisma.user.create({
-        data: {
-            name: 'Kelly',
-            email: 'kelly@prisma.io',
-        },
-        
-    })
-    console.log(user)
-}
-
-main()
-    .then(async () => {
-        await prisma.$disconnect()
-    })
-    .catch(async (e) => {
-        console.error(e)
-        await prisma.$disconnect()
-        process.exit(1)
-    })
