+++ src/app.module.ts
@@ -1,7 +1,10 @@
 import { Module } from '@nestjs/common';
 
+import { PrismaModule } from './prisma/prisma.module';
+import { UsersModule } from './users/users.module';
+
 @Module({
-  imports: [],
+  imports: [PrismaModule, UsersModule],
   controllers: [],
   providers: [],
 })
