+++ src/users/dto/create-user.dto.ts
@@ -1,10 +1,10 @@
 import { IsString, IsEmail, MaxLength } from 'class-validator';
 
 export class CreateUserDto {
-    @IsString()
-    @MaxLength(255)
-    name: string;
-  
-    @IsEmail()
-    email: string;
+  @IsString()
+  @MaxLength(255)
+  name: string;
+
+  @IsEmail()
+  email: string;
 }
